+++ MinerFull.java
@@ -2,7 +2,7 @@ import java.util.List;
 
 public class MinerFull extends Miner{
 
-    public MinerFull(String name, int resourceLimit, Point position, double rate, double animationRate){
+    public MinerFull(String name, int resourceLimit, Point position, int rate, int animationRate){
         super(name, resourceLimit, position, rate, animationRate, resourceLimit);
     }
